+++ extensions/gyazo-uploader/CHANGELOG.md
@@ -1,3 +1,9 @@
 # Gyazo Uploader Changelog
 
 ## [Initial Version] - 2023-08-12
+
+- initial release
+
+## [Remove old dependencies] - 2024-07-07
+
+- Use axios instead of gyazo-api library